Fix GPI-2 and enable TAGASPI
The rdma-core driver.h include is no longer installed:

 https://github.com/linux-rdma/rdma-core/commit/56dd87acd29279c81adfcf95f9d4d14806a2b615

So ibv_read_sysfs_file() is not defined. As the symbols is still
distributed, we simply add the missing prototype manually.

Similarly, the gaspi_get_system_mem() function is not available from the
gaspi public headers, so we define it in the max_mem.c test.

Fixes: https://jungle.bsc.es/git/rarias/bscpkgs/issues/7
Reviewed-by: Aleix Boné <abonerib@bsc.es>
Tested-by: Rodrigo Arias Mallo <rodrigo.arias@bsc.es>